+++ lib/process_hub.ex
@@ -172,7 +172,8 @@ defmodule ProcessHub do
           partition_tolerance_strategy:
             ProcessHub.Strategy.PartitionTolerance.Divergence.t()
             | ProcessHub.Strategy.PartitionTolerance.StaticQuorum.t()
-            | ProcessHub.Strategy.PartitionTolerance.DynamicQuorum.t(),
+            | ProcessHub.Strategy.PartitionTolerance.DynamicQuorum.t()
+            | ProcessHub.Strategy.PartitionTolerance.MajorityQuorum.t(),
           distribution_strategy:
             ProcessHub.Strategy.Distribution.ConsistentHashing.t()
             | ProcessHub.Strategy.Distribution.Guided.t()